Separated into files. Fixed placeholder.
